Data structures and Algorithms on graphs course

+++ Algorithmic_Toolbox/ded_course1_2020_05_28/week1_programming_challenges/2_maximum_pairwise_product/max_pairwise_product.cpp
@@ -15,9 +15,9 @@ int MaxPairwiseProduct(const std::vector<int>& numbers) {
 
     return max_product;
 }
-int MaxPairwiseProductF(const std::vector<int>& numbers) {
+long long MaxPairwiseProductF(const std::vector<int>& numbers) {
     int n = numbers.size();
-    int big1=0;int big2=0;
+    long long big1=0;long long big2=0;
     long long max_product;
     for(int i=0; i<n;i++){
         if(numbers[i]>big1){
